Generalized type system for scripting functions

+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/scripting/actions/types/WaitAction.kt
@@ -4,9 +4,9 @@ import ch.rmy.android.http_shortcuts.scripting.ExecutionContext
 import io.reactivex.Completable
 import java.util.concurrent.TimeUnit
 
-class WaitAction(private val duration: Long) : BaseAction() {
+class WaitAction(private val duration: Int) : BaseAction() {
 
     override fun execute(executionContext: ExecutionContext): Completable =
-        Completable.complete().delay(duration, TimeUnit.MILLISECONDS)
+        Completable.complete().delay(duration.toLong(), TimeUnit.MILLISECONDS)
 
 }
